modify authorization and views

+++ app/Http/Controllers/TagController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class TagController extends Controller
@@ -10,34 +11,51 @@ class TagController extends Controller
     use AuthorizesRequests;
 
     public function index(){
+        if (!Auth::user()->hasRole('admin') || !Auth::user()->hasPermissionTo('manage_tags')) {
+            abort(403, 'Unauthorized');
+        }
         $tags=Tag::all();
         return view('tags.index',compact('tags'));
     }
+
     public function create()
     {
+        if (!Auth::user()->hasRole('admin') || !Auth::user()->hasPermissionTo('manage_tags')) {
+            abort(403, 'Unauthorized');
+        }
         return view('tags.create');
     }
 
     function store(Request $request)
     {
+        if (!Auth::user()->hasRole('admin') || !Auth::user()->hasPermissionTo('manage_tags')) {
+            abort(403, 'Unauthorized');
+        }
 
         $validated = $request->validate([
             'name' => 'required|string|max:255|unique:tags,name',
         ]);
 
-        $tag = Tag::create($validated);
+        Tag::create($validated);
 
-        return response()->json(['tag' => $tag, 'status' => 'success']);
+        return redirect()->route('tags.index');
     }
     public function show(string $id)
     {
-        $tag=Tag::find($id);
-        return view('tags.show',compact('tag'));
+        if (!Auth::user()->hasRole('admin') || !Auth::user()->hasPermissionTo('manage_tags')) {
+            abort(403, 'Unauthorized');
+        }
+
+        $tags=Tag::find($id);
+        return view('tags.show',compact('tags'));
     }
 
-    public function edit(string $id)
+    public function edit(Tag $tag)
     {
-        return view('tags.edit');
+        if (!Auth::user()->hasRole('admin') || !Auth::user()->hasPermissionTo('manage_tags')) {
+            abort(403, 'Unauthorized');
+        }
+        return view('tags.edit',compact('tag'));
     }
 
     /**
@@ -45,6 +63,9 @@ class TagController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        if (!Auth::user()->hasRole('admin') || !Auth::user()->hasPermissionTo('manage_tags')) {
+            abort(403, 'Unauthorized');
+        }
         $validated=$request->validate([
             'name'=>'required|string|max:50',
         ]);
@@ -57,6 +78,9 @@ class TagController extends Controller
 
     public function destroy(string $id)
     {
+        if (!Auth::user()->hasRole('admin') || !Auth::user()->hasPermissionTo('manage_tags')) {
+            abort(403, 'Unauthorized');
+        }
         Tag::find($id)->delete();
         return redirect()->route('tags.index');
